+++ gulp/tasks/code-coverage.js
@@ -2,9 +2,11 @@ var gulp = require('gulp');
 var gutil = require('gulp-util');
 var helpers = require('../helpers');
 
-gulp.task('complexity', 'Run mocha tests', function(done) {
+gulp.task('code-coverage', 'Run mocha tests with coverage reporting turned on', function(done) {
   helpers.childProcess('node', [
-    'tasks/complexity.js'
+    '--harmony',
+    'tasks/mocha-runner.js',
+    '--coverage'
   ], function(error, stdout, stderr) {
     if(stderr) {
       gutil.log(gutil.colors.red(stderr));