+++ gulp/tasks/mocha.js
@@ -2,9 +2,10 @@ var gulp = require('gulp');
 var gutil = require('gulp-util');
 var helpers = require('../helpers');
 
-gulp.task('complexity', 'Run mocha tests', function(done) {
+gulp.task('mocha', 'Run mocha tests', function(done) {
   helpers.childProcess('node', [
-    'tasks/complexity.js'
+    '--harmony',
+    'tasks/mocha-runner.js'
   ], function(error, stdout, stderr) {
     if(stderr) {
       gutil.log(gutil.colors.red(stderr));