+++ web/app/application.jsx
@@ -1,3 +1,3 @@
 //load some modules that will load the startup code for our application
 require('./fluxe-loader');
-require('./components/core/router.js');
+require('./components/core/router.jsx');