+++ web/app/components/desktop/desktop.component.jsx
@@ -8,9 +8,11 @@ var Desktop = React.createClass({
     });
   },
   render: function() {
+    /* jshint ignore:start */
     return (
       <h1 id="test" className="test">Desktop</h1>
     );
+    /* jshint ignore:end */
   }
 });